+++ lib/core/routes/router.dart
@@ -1,11 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:interesting_places/core/screens/category_screen.dart';
-import 'package:interesting_places/core/screens/empty_route.dart';
 import 'package:interesting_places/core/screens/filter_screen.dart';
-import 'package:interesting_places/core/screens/main_screen.dart';
-import 'package:interesting_places/core/screens/place_list_map_screen.dart';
-import 'package:interesting_places/core/screens/place_point_map_screen.dart';
+import 'package:interesting_places/core/screens/map_screen.dart';
 import 'package:interesting_places/core/screens/new_place_screen.dart';
 import 'package:interesting_places/core/screens/place_details_screen.dart';
 import 'package:interesting_places/core/screens/place_list_screen.dart';
@@ -18,22 +15,8 @@ class AppRouter extends _$AppRouter {
   @override
   List<AutoRoute> get routes => [
         AutoRoute(
-          page: MainRoute.page,
+          page: PlaceListRoute.page,
           path: '/',
-          children: [
-            AutoRoute(
-              page: PlaceListRoute.page,
-              path: 'place_list',
-            ),
-            AutoRoute(
-              page: PlaceListMapRoute.page,
-              path: 'place_list_map',
-            ),
-            AutoRoute(
-              page: EmptyRoute.page,
-              path: 'empty',
-            ),
-          ],
         ),
         AutoRoute(
           page: NewPlaceRoute.page,
@@ -44,8 +27,8 @@ class AppRouter extends _$AppRouter {
           path: '/category',
         ),
         AutoRoute(
-          page: PlacePointMapRoute.page,
-          path: '/place_point_on_map',
+          page: MapRoute.page,
+          path: '/map',
         ),
         AutoRoute(
           page: FilterRoute.page,
