+++ lib/core/routes/router.gr.dart
@@ -21,22 +21,16 @@ abstract class _$AppRouter extends RootStackRouter {
         child: const CategoryScreen(),
       );
     },
-    EmptyRoute.name: (routeData) {
-      return AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const EmptyScreen(),
-      );
-    },
     FilterRoute.name: (routeData) {
       return AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const FilterScreen(),
       );
     },
-    MainRoute.name: (routeData) {
+    MapRoute.name: (routeData) {
       return AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const MainScreen(),
+        child: const MapScreen(),
       );
     },
     NewPlaceRoute.name: (routeData) {
@@ -55,22 +49,10 @@ abstract class _$AppRouter extends RootStackRouter {
         ),
       );
     },
-    PlaceListMapRoute.name: (routeData) {
-      return AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const PlaceListMapScreen(),
-      );
-    },
     PlaceListRoute.name: (routeData) {
       return AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const PlaceListScreen(),
-      );
-    },
-    PlacePointMapRoute.name: (routeData) {
-      return AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const PlacePointMapScreen(),
       );
     },
   };
@@ -91,20 +73,6 @@ class CategoryRoute extends PageRouteInfo<void> {
 }
 
 /// generated route for
-/// [EmptyScreen]
-class EmptyRoute extends PageRouteInfo<void> {
-  const EmptyRoute({List<PageRouteInfo>? children})
-      : super(
-          EmptyRoute.name,
-          initialChildren: children,
-        );
-
-  static const String name = 'EmptyRoute';
-
-  static const PageInfo<void> page = PageInfo<void>(name);
-}
-
-/// generated route for
 /// [FilterScreen]
 class FilterRoute extends PageRouteInfo<void> {
   const FilterRoute({List<PageRouteInfo>? children})
@@ -119,15 +87,15 @@ class FilterRoute extends PageRouteInfo<void> {
 }
 
 /// generated route for
-/// [MainScreen]
-class MainRoute extends PageRouteInfo<void> {
-  const MainRoute({List<PageRouteInfo>? children})
+/// [MapScreen]
+class MapRoute extends PageRouteInfo<void> {
+  const MapRoute({List<PageRouteInfo>? children})
       : super(
-          MainRoute.name,
+          MapRoute.name,
           initialChildren: children,
         );
 
-  static const String name = 'MainRoute';
+  static const String name = 'MapRoute';
 
   static const PageInfo<void> page = PageInfo<void>(name);
 }
@@ -185,20 +153,6 @@ class PlaceDetailsRouteArgs {
 }
 
 /// generated route for
-/// [PlaceListMapScreen]
-class PlaceListMapRoute extends PageRouteInfo<void> {
-  const PlaceListMapRoute({List<PageRouteInfo>? children})
-      : super(
-          PlaceListMapRoute.name,
-          initialChildren: children,
-        );
-
-  static const String name = 'PlaceListMapRoute';
-
-  static const PageInfo<void> page = PageInfo<void>(name);
-}
-
-/// generated route for
 /// [PlaceListScreen]
 class PlaceListRoute extends PageRouteInfo<void> {
   const PlaceListRoute({List<PageRouteInfo>? children})
@@ -208,20 +162,6 @@ class PlaceListRoute extends PageRouteInfo<void> {
         );
 
   static const String name = 'PlaceListRoute';
-
-  static const PageInfo<void> page = PageInfo<void>(name);
-}
-
-/// generated route for
-/// [PlacePointMapScreen]
-class PlacePointMapRoute extends PageRouteInfo<void> {
-  const PlacePointMapRoute({List<PageRouteInfo>? children})
-      : super(
-          PlacePointMapRoute.name,
-          initialChildren: children,
-        );
-
-  static const String name = 'PlacePointMapRoute';
 
   static const PageInfo<void> page = PageInfo<void>(name);
 }
