--- lib/core/screens/empty_route.dart
@@ -1,12 +1,0 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:flutter/material.dart';
-
-@RoutePage()
-class EmptyScreen extends StatelessWidget {
-  const EmptyScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox.shrink();
-  }
-}
