+++ lib/features/place_list/domain/entity/place_entity.dart
@@ -3,7 +3,6 @@ import 'package:interesting_places/core/data/models/category.dart';
 
 class PlaceEntity {
   const PlaceEntity({
-    required this.id,
     required this.imageList,
     required this.name,
     required this.description,
@@ -13,7 +12,6 @@ class PlaceEntity {
     this.distance,
   });
 
-  final int id;
   final List<Uint8List> imageList;
   final String name;
   final String description;
